+++ tensorrt_llm/models/gemma/smoothquant.py
@@ -388,11 +388,11 @@ class LlamaAttentionExtend(LlamaAttention):
         self,
         hidden_states: torch.Tensor,
         attention_mask: Optional[torch.Tensor] = None,
-        position_ids: Optional[torch.IntTensor] = None,
+        position_ids: Optional[torch.LongTensor] = None,
         past_key_value: "Optional[Cache]" = None,
         output_attentions: bool = False,
         use_cache: bool = False,
-        cache_position: Optional[torch.IntTensor] = None,
+        cache_position: Optional[torch.LongTensor] = None,
         **kwargs,
     ) -> Tuple[torch.Tensor, Optional[torch.Tensor],
                Optional[Tuple[torch.Tensor]]]:
